Change `content` parameter of `file` resource
Content can now be populated with placeholders that are replaced by
the output of commands that run on the client.
This way file content can now be substituted by variables that are
known ahead of time on the server side, as well a dynamic output from
custom commands on the client side.