Add `when` parameter to triggers
This parameter ensures that an `execute` resource is only triggered by
another resource if its state after being applied matches one of the
states in `when`.
This enables users to trigger different `execute` resource based on
the state of the triggering resource, e.g. `created` or `deleted`.
By default (when the optional `when` array is missing) an `execute`
resource is triggered when a resource is `created`, `changed` and `deleted`.